[40기 곽민경] BestProductList 오류 수정
