+++ src/pages/Main/Main.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Promotion from './MainList/Promotion';
-import BestProductList from './MainList/BestProductList';
+import BestProduct from './MainList/BestProduct';
 import Delivery from './MainList/Delivery';
 import './Main.scss';
 
@@ -9,7 +9,7 @@ const Main = () => {
     <div className="main">
       <Promotion />
       <Delivery />
-      <BestProductList />
+      <BestProduct />
     </div>
   );
 };
